add BacktestExecutor and extract method.
